Scroll marker color changed
originally, when used with the highlight scroll marker package, it was hard to tell which marker was which. the orange and yellow are hard to tell apart. now, I made the marker for this one to be a bright blue, so it stands out form the red and orange of the other marker packages.